用户加入token，保存监管数据修改

+++ shianxian-supervise/src/main/java/cn/shianxian/supervise/info/pojo/SuperviseInfoSub.java
@@ -68,7 +68,7 @@ public class SuperviseInfoSub {
      * 结果分值
      */
     @Column(name = "score")
-    private Long score;
+    private Double score;
 
     /**
      * 整改意见
